Refactor SocialGroups component
Refactor SocialGroups.tsx into smaller, more manageable files.
