+++ src/components/social/SocialGroups.tsx
@@ -1,39 +1,31 @@
 
 import { useState, useEffect } from "react";
 import { Button } from "@/components/ui/button";
-import { Card, CardHeader, CardContent, CardFooter } from "@/components/ui/card";
-import { Badge } from "@/components/ui/badge";
+import { Card, CardContent } from "@/components/ui/card";
 import { toast } from "sonner";
-import { Users, MapPin, PlusCircle, AlertCircle, ThumbsUp, MessageSquare } from "lucide-react";
+import { PlusCircle } from "lucide-react";
 import { supabase } from "@/integrations/supabase";
 import { useAuth } from "@/context/AuthContext";
-import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogFooter } from "@/components/ui/dialog";
-import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs";
-import { Avatar } from "@/components/ui/avatar";
-import { Textarea } from "@/components/ui/textarea";
-import { Input } from "@/components/ui/input";
 import { SocialGroup, SocialPost } from "./types";
 import { useSocialPosts } from "@/hooks/useSocialPosts";
+import GroupCard from "./groups/GroupCard";
+import GroupDetailView from "./groups/GroupDetailView";
+import CreateGroupForm from "./groups/CreateGroupForm";
 
 export default function SocialGroups() {
   const [groups, setGroups] = useState<SocialGroup[]>([]);
   const [recommendedGroups, setRecommendedGroups] = useState<SocialGroup[]>([]);
   const [selectedGroup, setSelectedGroup] = useState<SocialGroup | null>(null);
-  const [activeTabId, setActiveTabId] = useState<string | null>(null);
   const [groupPosts, setGroupPosts] = useState<SocialPost[]>([]);
   const [pendingPosts, setPendingPosts] = useState<SocialPost[]>([]);
   const [newGroupPost, setNewGroupPost] = useState("");
   const [joinedGroups, setJoinedGroups] = useState<string[]>([]);
   const [isLoading, setIsLoading] = useState(true);
-
   const [showCreateModal, setShowCreateModal] = useState(false);
-  const [newGroupName, setNewGroupName] = useState("");
-  const [newGroupDescription, setNewGroupDescription] = useState("");
 
   const { user } = useAuth();
-  const { createPost, votePost } = useSocialPosts();
+  const { createPost } = useSocialPosts();
   
-  // Mock moderatePost and isSubmitting functions for now
   const moderatePost = async (postId: string, approve: boolean) => {
     console.log("Moderating post:", postId, approve);
     return true;
@@ -214,35 +206,9 @@ export default function SocialGroups() {
       toast.error("Something went wrong");
     }
   };
-
-  const handleCreateGroup = async () => {
-    if (!user) {
-      toast.error("You must be logged in to create a group");
-      return;
-    }
-
-    if (!newGroupName.trim()) {
-      toast.error("Group name cannot be empty");
-      return;
-    }
-
-    try {
-      console.log("Attempting to create group:", { name: newGroupName, description: newGroupDescription, owner_id: user.id });
-      toast.success(`Group "${newGroupName}" created successfully!`);
-
-      setNewGroupName("");
-      setNewGroupDescription("");
-      setShowCreateModal(false);
-      fetchGroups();
-    } catch (err) {
-      console.error("Error creating group:", err);
-      toast.error(`Failed to create group: ${err.message}`);
-    }
-  };
   
   const handleSelectGroup = (group: SocialGroup) => {
     setSelectedGroup(group);
-    setActiveTabId(group.id.toString());
     setNewGroupPost("");
   };
   
@@ -262,220 +228,23 @@ export default function SocialGroups() {
       fetchGroupPosts(selectedGroup.id.toString());
     }
   };
-  
-  const getActivityBadge = (activityLevel: string) => {
-    switch (activityLevel) {
-      case 'active':
-        return <Badge className="bg-green-500">Active</Badge>;
-      case 'new':
-        return <Badge className="bg-blue-500">New</Badge>;
-      case 'quiet':
-        return <Badge className="bg-gray-500">Quiet</Badge>;
-      default:
-        return null;
-    }
-  };
 
   return (
     <div className="space-y-8">
       {selectedGroup ? (
-        <div>
-          {/* Group header with back button */}
-          <div className="flex items-center gap-4 mb-6">
-            <Button 
-              variant="outline" 
-              onClick={() => {
-                setSelectedGroup(null);
-                setActiveTabId(null);
-              }}
-            >
-              Back to Groups
-            </Button>
-            <h2 className="text-xl font-semibold">{selectedGroup.name}</h2>
-            {selectedGroup.isAdmin && (
-              <Badge className="bg-purple-500">Admin</Badge>
-            )}
-          </div>
-          
-          {/* Group cover and details */}
-          <Card className="mb-6">
-            <div className="h-40 overflow-hidden">
-              <img 
-                src={selectedGroup.cover} 
-                alt={selectedGroup.name} 
-                className="w-full h-full object-cover"
-              />
-            </div>
-            <CardContent className="p-6">
-              <div className="flex justify-between items-start mb-4">
-                <div>
-                  <h3 className="text-lg font-semibold">{selectedGroup.name}</h3>
-                  <div className="flex items-center text-sm text-gray-600 mt-1">
-                    <Users size={16} className="mr-1" /> {selectedGroup.members} members
-                  </div>
-                  {selectedGroup.location && (
-                    <div className="flex items-center text-sm text-gray-600 mt-1">
-                      <MapPin size={16} className="mr-1" /> {selectedGroup.location}
-                    </div>
-                  )}
-                </div>
-                {getActivityBadge(selectedGroup.activityLevel)}
-              </div>
-              <p className="text-gray-700">{selectedGroup.description}</p>
-              
-              {!joinedGroups.includes(selectedGroup.id.toString()) && !selectedGroup.isAdmin && (
-                <Button
-                  className="mt-4"
-                  onClick={() => handleJoinGroup(selectedGroup.id.toString())}
-                >
-                  Join Group
-                </Button>
-              )}
-            </CardContent>
-          </Card>
-          
-          {/* Group posts section with tabs */}
-          <div className="space-y-6">
-            {/* New post creation */}
-            {(joinedGroups.includes(selectedGroup.id.toString()) || selectedGroup.isAdmin) && (
-              <Card className="border-2 border-blue-100 mb-6">
-                <CardContent className="pt-6">
-                  <h3 className="text-lg font-semibold mb-3">Create a new post</h3>
-                  <Textarea 
-                    value={newGroupPost} 
-                    onChange={(e) => setNewGroupPost(e.target.value)}
-                    placeholder="Share something with the group..."
-                    className="min-h-[120px] mb-4"
-                    disabled={isSubmitting}
-                  />
-                  <div className="flex justify-end">
-                    <Button
-                      onClick={handlePostSubmit}
-                      disabled={isSubmitting || !newGroupPost.trim()}
-                    >
-                      {isSubmitting ? "Submitting..." : "Post to Group"}
-                    </Button>
-                  </div>
-                  <div className="text-xs text-gray-500 mt-2">
-                    Group posts require approval from the group admin
-                  </div>
-                </CardContent>
-              </Card>
-            )}
-            
-            {/* Admin moderation section */}
-            {selectedGroup.isAdmin && pendingPosts.length > 0 && (
-              <div className="mb-6">
-                <h3 className="text-lg font-semibold mb-4">Posts Awaiting Approval</h3>
-                <div className="space-y-4">
-                  {pendingPosts.map((post) => (
-                    <Card key={post.id} className="border-2 border-amber-200">
-                      <CardHeader className="flex flex-row items-center gap-4 pb-2">
-                        <Avatar className="w-10 h-10">
-                          <img src={post.authorAvatar} alt={post.author} />
-                        </Avatar>
-                        <div className="flex flex-col">
-                          <div className="flex items-center gap-2">
-                            <h4 className="font-semibold">{post.author}</h4>
-                            <Badge className="bg-amber-500">Pending</Badge>
-                          </div>
-                          <p className="text-sm text-gray-500">{post.date}</p>
-                        </div>
-                      </CardHeader>
-                      <CardContent>
-                        <p className="whitespace-pre-line">{post.content}</p>
-                        {post.image && (
-                          <div className="mt-4">
-                            <img 
-                              src={post.image} 
-                              alt="Post image" 
-                              className="rounded-md max-h-[300px] object-cover" 
-                            />
-                          </div>
-                        )}
-                      </CardContent>
-                      <CardFooter className="border-t pt-4">
-                        <div className="flex gap-4 w-full">
-                          <Button 
-                            variant="default"
-                            className="bg-green-600 hover:bg-green-700" 
-                            onClick={() => handleModeratePost(post.id.toString(), true)}
-                          >
-                            Approve
-                          </Button>
-                          <Button 
-                            variant="outline"
-                            className="text-red-600 border-red-600 hover:bg-red-50" 
-                            onClick={() => handleModeratePost(post.id.toString(), false)}
-                          >
-                            Reject
-                          </Button>
-                        </div>
-                      </CardFooter>
-                    </Card>
-                  ))}
-                </div>
-              </div>
-            )}
-            
-            {/* Group posts */}
-            <div>
-              <h3 className="text-lg font-semibold mb-4">Group Posts</h3>
-              {groupPosts.length === 0 ? (
-                <Card className="text-center py-8">
-                  <CardContent>
-                    {joinedGroups.includes(selectedGroup.id.toString()) || selectedGroup.isAdmin ? (
-                      <p className="text-gray-500">No posts yet. Be the first to share!</p>
-                    ) : (
-                      <p className="text-gray-500">Join this group to see and create posts</p>
-                    )}
-                  </CardContent>
-                </Card>
-              ) : (
-                <div className="space-y-4">
-                  {groupPosts.map((post) => (
-                    <Card key={post.id}>
-                      <CardHeader className="flex flex-row items-center gap-4 pb-2">
-                        <Avatar className="w-10 h-10">
-                          <img src={post.authorAvatar} alt={post.author} />
-                        </Avatar>
-                        <div className="flex flex-col">
-                          <div className="flex items-center gap-2">
-                            <h4 className="font-semibold">{post.author}</h4>
-                            {post.isOwnPost && <Badge variant="outline">You</Badge>}
-                          </div>
-                          <p className="text-sm text-gray-500">{post.date}</p>
-                        </div>
-                      </CardHeader>
-                      <CardContent>
-                        <p className="whitespace-pre-line">{post.content}</p>
-                        {post.image && (
-                          <div className="mt-4">
-                            <img 
-                              src={post.image} 
-                              alt="Post image" 
-                              className="rounded-md max-h-[300px] object-cover" 
-                            />
-                          </div>
-                        )}
-                      </CardContent>
-                      <CardFooter className="border-t pt-4">
-                        <div className="flex gap-6 w-full">
-                          <Button variant="ghost" size="sm" className="flex items-center gap-1">
-                            <ThumbsUp size={18} /> {post.likes}
-                          </Button>
-                          <Button variant="ghost" size="sm" className="flex items-center gap-1">
-                            <MessageSquare size={18} /> {post.comments}
-                          </Button>
-                        </div>
-                      </CardFooter>
-                    </Card>
-                  ))}
-                </div>
-              )}
-            </div>
-          </div>
-        </div>
+        <GroupDetailView
+          group={selectedGroup}
+          isJoined={joinedGroups.includes(selectedGroup.id.toString())}
+          groupPosts={groupPosts}
+          pendingPosts={pendingPosts}
+          newGroupPost={newGroupPost}
+          setNewGroupPost={setNewGroupPost}
+          onBack={() => setSelectedGroup(null)}
+          onJoinGroup={handleJoinGroup}
+          onPostSubmit={handlePostSubmit}
+          onModeratePost={handleModeratePost}
+          isSubmitting={isSubmitting}
+        />
       ) : (
         <>
           {/* Pam's recommendations */}
@@ -509,55 +278,13 @@ export default function SocialGroups() {
             ) : (
               <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
                 {groups.map((group) => (
-                  <Card key={group.id} className="overflow-hidden flex flex-col">
-                    <div className="h-40 overflow-hidden cursor-pointer" onClick={() => handleSelectGroup(group)}>
-                      <img 
-                        src={group.cover} 
-                        alt={group.name} 
-                        className="w-full h-full object-cover"
-                      />
-                    </div>
-                    <CardHeader className="pb-2">
-                      <div className="flex justify-between items-start">
-                        <h4 className="font-semibold text-lg cursor-pointer" onClick={() => handleSelectGroup(group)}>
-                          {group.name}
-                        </h4>
-                        {getActivityBadge(group.activityLevel)}
-                      </div>
-                      <div className="flex items-center text-sm text-gray-600">
-                        <Users size={16} className="mr-1" /> {group.members} members
-                      </div>
-                      {group.location && (
-                        <div className="flex items-center text-sm text-gray-600">
-                          <MapPin size={16} className="mr-1" /> {group.location}
-                        </div>
-                      )}
-                      {group.isAdmin && (
-                        <Badge className="bg-purple-500 mt-1">Admin</Badge>
-                      )}
-                    </CardHeader>
-                    <CardContent className="pb-2 flex-grow">
-                      <p className="text-sm text-gray-700 line-clamp-2">{group.description}</p>
-                    </CardContent>
-                    <CardFooter className="pt-2 flex justify-between">
-                      <Button 
-                        variant="default"
-                        className="flex-1 mr-2"
-                        onClick={() => handleSelectGroup(group)}
-                      >
-                        View Group
-                      </Button>
-                      {!group.isAdmin && (
-                        <Button 
-                          variant={joinedGroups.includes(group.id.toString()) ? "outline" : "secondary"} 
-                          className="flex-1 ml-2"
-                          onClick={() => handleJoinGroup(group.id.toString())}
-                        >
-                          {joinedGroups.includes(group.id.toString()) ? "Leave" : "Join"}
-                        </Button>
-                      )}
-                    </CardFooter>
-                  </Card>
+                  <GroupCard
+                    key={group.id}
+                    group={group}
+                    isJoined={joinedGroups.includes(group.id.toString())}
+                    onSelectGroup={handleSelectGroup}
+                    onJoinGroup={handleJoinGroup}
+                  />
                 ))}
               </div>
             )}
@@ -579,39 +306,11 @@ export default function SocialGroups() {
       )}
       
       {/* Create Group Modal */}
-      <Dialog open={showCreateModal} onOpenChange={setShowCreateModal}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Create New Group</DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4">
-            <div>
-              <label className="text-sm font-medium">Group Name</label>
-              <Input
-                value={newGroupName}
-                onChange={(e) => setNewGroupName(e.target.value)}
-                placeholder="Enter group name"
-              />
-            </div>
-            <div>
-              <label className="text-sm font-medium">Description</label>
-              <Textarea
-                value={newGroupDescription}
-                onChange={(e) => setNewGroupDescription(e.target.value)}
-                placeholder="Describe your group"
-              />
-            </div>
-          </div>
-          <DialogFooter>
-            <Button variant="outline" onClick={() => setShowCreateModal(false)}>
-              Cancel
-            </Button>
-            <Button onClick={handleCreateGroup}>
-              Create Group
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+      <CreateGroupForm
+        isOpen={showCreateModal}
+        onClose={() => setShowCreateModal(false)}
+        onGroupCreated={fetchGroups}
+      />
     </div>
   );
 }
